+++ packages/router/test/lib/helpers/prepare.spec.ts
@@ -1,4 +1,6 @@
 import { expect, jsonifyError } from "@connext/nxtp-utils";
+import { BigNumber } from "ethers";
+import { parseEther, parseUnits } from "ethers/lib/utils";
 import { AmountInvalid } from "../../../src/lib/errors/prepare";
 import * as PrepareHelperFns from "../../../src/lib/helpers/prepare";
 import * as SharedHelperFns from "../../../src/lib/helpers/shared";
@@ -62,6 +64,21 @@ describe("getReceiverAmount", () => {
     } catch (e) {
       expect(e.message).to.be.eq(err.message);
     }
+  });
+
+  it("should work for 6 to 18", async () => {
+    const result = await getReceiverAmount(parseUnits("1", 6).toString(), 6, 18);
+    expect(result).to.be.eq(parseEther("0.9995").toString());
+  });
+
+  it("should work for 18 to 6", async () => {
+    const result = await getReceiverAmount(parseUnits("1", 18).toString(), 18, 6);
+    expect(result).to.be.eq(parseUnits("0.9995", 6).toString());
+  });
+
+  it("should work for decimals", async () => {
+    const result = await getReceiverAmount("104731245", 6, 6);
+    expect(() => BigNumber.from(result)).to.not.throw();
   });
 
   it("should fail if its a decimal string", async () => {