+++ packages/sdk/src/sdk.ts
@@ -1,4 +1,4 @@
-import { providers, Signer, utils, BigNumber } from "ethers";
+import { providers, Signer, utils } from "ethers";
 import { Evt } from "evt";
 import {
   UserNxtpNatsMessagingService,
@@ -9,8 +9,6 @@ import {
   Logger,
   createLoggingContext,
   TransactionData,
-  RequestContext,
-  MethodContext,
 } from "@connext/nxtp-utils";
 
 import { getDeployedChainIdsForGasFee } from "./transactionManager/transactionManager";
@@ -93,6 +91,7 @@ export class NxtpSdk {
         };
       };
       signer: Signer;
+      messagingSigner?: Signer;
       logger?: Logger;
       network?: "testnet" | "mainnet" | "local";
       natsUrl?: string;
@@ -102,7 +101,8 @@ export class NxtpSdk {
       sdkBase?: NxtpSdkBase;
     },
   ) {
-    const { chainConfig, signer, messaging, natsUrl, authUrl, logger, network, skipPolling, sdkBase } = this.config;
+    const { chainConfig, signer, messagingSigner, messaging, natsUrl, authUrl, logger, network, skipPolling, sdkBase } =
+      this.config;
 
     this.logger = logger ?? new Logger({ name: "NxtpSdk" });
 
@@ -117,6 +117,7 @@ export class NxtpSdk {
         signer,
         logger: this.logger.child({ name: "NxtpSdkBase" }),
         network,
+        messagingSigner,
         skipPolling,
       });
   }
@@ -150,6 +151,11 @@ export class NxtpSdk {
    */
   public async getHistoricalTransactions(): Promise<HistoricalTransaction[]> {
     return this.sdkBase.getHistoricalTransactions();
+  }
+
+  public async estimateFulfillFee(txData: TransactionData, signatureForFee: string, relayerFee: string) {
+    const { requestContext, methodContext } = createLoggingContext("estimateFulfillFee");
+    return this.sdkBase.estimateFulfillFee(txData, signatureForFee, relayerFee, requestContext, methodContext);
   }
 
   /**
@@ -319,7 +325,8 @@ export class NxtpSdk {
     let calculateRelayerFee = relayerFee;
     const chainIdsForPriceOracle = getDeployedChainIdsForGasFee();
     if (useRelayers && chainIdsForPriceOracle.includes(txData.receivingChainId)) {
-      const gasNeeded = await this.estimateFulfillFee(txData, relayerFee, requestContext, methodContext);
+      const gasNeeded = await this.sdkBase.estimateFulfillFee(txData, "0x", "0", requestContext, methodContext);
+
       this.logger.info(
         `Calculating Gas Fee for fulfill tx. neededGas = ${gasNeeded.toString()}`,
         requestContext,
@@ -329,7 +336,7 @@ export class NxtpSdk {
       calculateRelayerFee = gasNeeded.toString();
     }
 
-    this.logger.info("Generating fulfill signature", requestContext, methodContext);
+    this.logger.info("Generating fulfill signature", requestContext, methodContext, { calculateRelayerFee });
     const signature = await signFulfillTransactionPayload(
       txData.transactionId,
       calculateRelayerFee,
@@ -349,7 +356,7 @@ export class NxtpSdk {
         throw new EncryptionError("decryption failed", jsonifyError(e));
       }
     }
-    const response = await this.sdkBase.fulfillTransfer(params, signature, callData, relayerFee, useRelayers);
+    const response = await this.sdkBase.fulfillTransfer(params, signature, callData, calculateRelayerFee, useRelayers);
 
     if (useRelayers) {
       return { metaTxResponse: response.metaTxResponse };
@@ -362,30 +369,6 @@ export class NxtpSdk {
     }
   }
 
-  public async estimateFulfillFee(
-    txData: TransactionData,
-    relayerFee: string,
-    requestContext: RequestContext,
-    methodContext: MethodContext,
-  ): Promise<BigNumber> {
-    const signatureForFee = await signFulfillTransactionPayload(
-      txData.transactionId,
-      relayerFee,
-      txData.receivingChainId,
-      txData.receivingChainTxManagerAddress,
-      this.config.signer,
-    );
-
-    const estimateFulfillFeeResponse = await this.sdkBase.estimateFulfillFee(
-      txData,
-      signatureForFee,
-      relayerFee,
-      requestContext,
-      methodContext,
-    );
-    return estimateFulfillFeeResponse;
-  }
-
   /**
    * Cancels the given transaction
    *
@@ -396,7 +379,6 @@ export class NxtpSdk {
    * @param chainId - Chain to cancel the transaction on
    * @returns A TransactionResponse when the transaction was submitted, not mined
    */
-
   public async cancel(cancelParams: CancelParams, chainId: number): Promise<providers.TransactionResponse> {
     const { requestContext, methodContext } = createLoggingContext(
       this.cancel.name,